+++ AirLib/include/vehicles/multirotor/RotorParams.hpp
@@ -50,11 +50,11 @@ namespace msr {
             void calculateMaxThrust() {
                 revolutions_per_second = max_rpm / 60;
                 max_speed = revolutions_per_second * 2 * M_PIf;  // radians / sec
-                max_speed_square = pow(max_speed, 2.0f);
+                max_speed_square = powf(max_speed, 2.0f);
 
                 real_T nsquared = revolutions_per_second * revolutions_per_second;
-                max_thrust = C_T * air_density * nsquared * pow(propeller_diameter, 4);
-                max_torque = C_P * air_density * nsquared * pow(propeller_diameter, 5) / (2 * M_PIf);
+                max_thrust = C_T * air_density * nsquared * powf(propeller_diameter, 4);
+                max_torque = C_P * air_density * nsquared * powf(propeller_diameter, 5) / (2 * M_PIf);
             }
 
         };
